Chore: move tag length constants out to constants file
There was a circular dependency going on which prevented these values
from ever being validated.
