+++ src/lib/services/tag-schema.ts
@@ -1,9 +1,8 @@
 import Joi from 'joi';
 
 import { customJoi } from '../routes/util';
+import { TAG_MAX_LENGTH, TAG_MIN_LENGTH } from '../util';
 
-export const TAG_MIN_LENGTH = 2;
-export const TAG_MAX_LENGTH = 50;
 export const tagSchema = Joi.object()
     .keys({
         value: Joi.string().min(TAG_MIN_LENGTH).max(TAG_MAX_LENGTH),
