+++ src/lib/util/constants.ts
@@ -17,6 +17,9 @@ export const PREDEFINED_ROLE_TYPES = [ROOT_ROLE_TYPE, PROJECT_ROLE_TYPE];
 export const ROOT_ROLE_TYPES = [ROOT_ROLE_TYPE, CUSTOM_ROOT_ROLE_TYPE];
 export const PROJECT_ROLE_TYPES = [PROJECT_ROLE_TYPE, CUSTOM_PROJECT_ROLE_TYPE];
 
+export const TAG_MIN_LENGTH = 2;
+export const TAG_MAX_LENGTH = 50;
+
 /* CONTEXT FIELD OPERATORS */
 
 export const NOT_IN = 'NOT_IN';
